backend/requirements/base.txt
fastapi==0.115.7
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20

python-socketio==5.13.0
python-jose==3.4.0
passlib[bcrypt]==1.7.4

requests==2.32.3
aiohttp==3.11.11
async-timeout
aiocache
aiofiles

sqlalchemy==2.0.38
alembic==1.14.0
peewee==3.17.9
peewee-migrate==1.12.2
psycopg2-binary==2.9.9
PyMySQL==1.1.1
bcrypt==4.3.0

pymongo
redis

argon2-cffi==23.1.0
APScheduler==3.10.4

RestrictedPython==8.0

loguru==0.7.2
asgiref==3.8.1

openai
anthropic
google-generativeai==0.8.4
tiktoken

langchain==0.3.19
langchain-community==0.3.18

fake-useragent==2.1.0
chromadb==0.6.3

ftfy==6.2.3
pypdf==4.3.1
fpdf2==2.8.2
pymdown-extensions==10.14.2
docx2txt==0.8
Markdown==3.7
validators==0.34.0
rank-bm25==0.2.2

PyJWT[crypto]==2.10.1
cryptography>=43.0.0
authlib==1.4.1

youtube-transcript-api==1.0.3
pydub
duckduckgo-search~=8.0.0

ldap3==2.9.1

tencentcloud-sdk-python==3.0.1336
